try to fix test
